// src/generated/model/logg.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Logg {
    level?: string;
    message?: string;
    jsFileUrl?: string;
    lineNumber?: string;
    columnNumber?: string;
    url?: string;
    userAgent?: string;
}